pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ppde"
version = "0.1.0"
description = "Personalized Predictive Debugging Engine"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1.0",
]

[project.scripts]
ppde = "ppde.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ppde*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B"]
